Chunk server: fix disk IO handles leak in buffering layer by deleting handles in completion handler. Do not allow to close unstable object store blocks by "emergency" out of file descriptors scans with shorter LRU expiration, keep those around for at least one lease interval to allow client to handle communication errors by "re-allocating" object store block. Make object store block hash prefix generation byte order independent. Do not attempt to load stable object block header when handling make chunk stable with unspecified chunk size, return an error instead, as make stable issued by the meta server in response to the chunk server orderly lease relinquish must have chunk size. Do not cleanup object store block in allocation handling, by invoking delete chunk, as later will all the object store block to pending delete queue, and processing of this queue cannot reliably determine stale deletes as object store blocks table entries, unlike chunk table entries, are not permanent, and might be evicted and loaded on demand. Client library: with object store files close even a single chunk writer as soon as chunk is complete / full as re-write is not supported, in order to minimize the number of non-stable chunks, and the corresponding memory buffers.

+++ src/cc/chunk/KfsOps.cc
@@ -1220,7 +1220,7 @@ AllocChunkOp::Execute()
     // Allocation implicitly invalidates all previously existed write leases.
     gLeaseClerk.UnRegisterLease(chunkId, chunkVersion);
     mustExistFlag = mustExistFlag || 1 < chunkVersion;
-    if (! mustExistFlag) {
+    if (! mustExistFlag && 0 <= chunkVersion) {
         const int ret = gChunkManager.DeleteChunk(chunkId, chunkVersion);
         if (ret != -EBADF) {
             KFS_LOG_STREAM_WARN <<
